# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DOGFOOD 2026 Hackathon Management & Judging Platform API
  description: Self-hostable, offline-capable REST API for hackathon lifecycles, judging, score normalization, community voting, and exports.
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
    description: Local Backend Server
paths:
...
  /events/{eventId}/judges/assign/auto:
    post:
      summary: Execute deterministic judge assignment
      tags: [Judging]
      security:
        - CookieAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                seed: { type: integer, default: 42 }
      responses:
        '200':
          description: Assignments generated with workload statistics
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: dogfood_session
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT